+++ src/mode.rs
@@ -24,32 +24,21 @@ impl RustfixMode {
 pub enum Mode {
     /// The test passes a full execution of the rustc driver
     Pass,
-    /// The test produces an executable binary that can get executed on the host
-    Run {
-        /// The expected exit code
-        exit_code: i32,
-    },
     /// The rustc driver panicked
     Panic,
     /// The rustc driver emitted an error
     Fail {
         /// Whether failing tests must have error patterns. Set to false if you just care about .stderr output.
         require_patterns: bool,
-        /// When to run rustfix on the test
-        rustfix: RustfixMode,
     },
     /// Run the tests, but always pass them as long as all annotations are satisfied and stderr files match.
-    Yolo {
-        /// When to run rustfix on the test
-        rustfix: RustfixMode,
-    },
+    Yolo,
 }
 
 impl Mode {
     #[allow(clippy::result_large_err)]
     pub(crate) fn ok(self, status: ExitStatus) -> Result<(), Error> {
         let expected = match self {
-            Mode::Run { exit_code } => exit_code,
             Mode::Pass => 0,
             Mode::Panic => 101,
             Mode::Fail { .. } => 1,
@@ -59,7 +48,7 @@ impl Mode {
             Ok(())
         } else {
             Err(Error::ExitStatus {
-                mode: self,
+                mode: self.to_string(),
                 status,
                 expected,
             })
@@ -70,14 +59,12 @@ impl Mode {
 impl Display for Mode {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Mode::Run { exit_code } => write!(f, "run({exit_code})"),
             Mode::Pass => write!(f, "pass"),
             Mode::Panic => write!(f, "panic"),
             Mode::Fail {
                 require_patterns: _,
-                rustfix: _,
             } => write!(f, "fail"),
-            Mode::Yolo { rustfix: _ } => write!(f, "yolo"),
+            Mode::Yolo => write!(f, "yolo"),
         }
     }
 }